successfully updated unit_ordered reset function

+++ schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE products
     item_id INT(100) AUTO_INCREMENT NOT NULL,
     product_name VARCHAR(100) NULL,
     department_name VARCHAR(100) NULL,
-    price DECIMAL(10,2) NULL,
+    price DOUBLE(10,2) NULL,
     stock_quantity INT(100) NULL,
     units_ordered INT(100) NULL,
     PRIMARY KEY (item_id)
